refactor(scanner): moved the error state & animation to scanner container
re #1501
